MI-314: Fixed linting issues & rename folder
