+++ packages/constructs/nodejs-function-from-entry/.npmignore
