+++ packages/constructs/nodejs-function-from-entry/index.ts
@@ -1,6 +1,6 @@
 import {
   NodejsFunctionFromEntry,
   NodejsFunctionFromEntryProps,
-} from "./lib/nodejs-function-from-entry-construct";
+} from "./lib/nodejs-function-from-entry";
 
 export { NodejsFunctionFromEntry, NodejsFunctionFromEntryProps };
