+++ packages/constructs/nodejs-function-from-entry/lib/nodejs-function-from-entry.ts
