+++ packages/constructs/nodejs-function-from-entry/project.json
@@ -1,14 +1,14 @@
 {
-  "name": "nodejs-function-from-entry-construct",
+  "name": "nodejs-function-from-entry",
   "$schema": "../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "packages/constructs/nodejs-function-from-entry-construct/lib",
+  "sourceRoot": "packages/constructs/nodejs-function-from-entry/lib",
   "projectType": "application",
   "targets": {
     "build": {
       "executor": "nx:run-commands",
       "options": {
         "command": "tsc --project tsconfig.app.json",
-        "cwd": "packages/constructs/nodejs-function-from-entry-construct"
+        "cwd": "packages/constructs/nodejs-function-from-entry"
       }
     },
     "lint": {
@@ -19,7 +19,7 @@
       "executor": "@nx/jest:jest",
       "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
       "options": {
-        "jestConfig": "packages/constructs/nodejs-function-from-entry-construct/jest.config.ts",
+        "jestConfig": "packages/constructs/nodejs-function-from-entry/jest.config.ts",
         "passWithNoTests": true
       }
     },
@@ -27,7 +27,7 @@
       "executor": "nx:run-commands",
       "options": {
         "command": "npm publish --access public",
-        "cwd": "packages/constructs/nodejs-function-from-entry-construct"
+        "cwd": "packages/constructs/nodejs-function-from-entry"
       },
       "dependsOn": ["build"]
     }
