+++ packages/constructs/nodejs-function-from-entry/tsconfig.app.json
