+++ packages/constructs/nodejs-function-from-entry/tsconfig.json
