+++ packages/constructs/nodejs-function-from-entry/tsconfig.spec.json
